Loaded Algorithm from Abstract class
